pyproject.toml
[tool.poetry]
name = "travel-shop"
version = "0.0.1"
description = ""
authors = ["Vovak <Vovakjh@yandex.ru>"]
readme = "README.md"
packages = [{include = "travel_shop"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
faker = "^19.3.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-bugbear = "^23.6.5"
flake8-simplify = "^0.20.0"
flake8-class-attributes-order = "^0.1.3"
flake8-clean-block = "^0.1.2"
flake8-commas = "^2.1.0"
flake8-use-fstring = "^1.4"
flake8-use-pathlib = "^0.3.0"
flake8-absolute-import = "^1.0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
